requirements.txt
fastapi 
uvicorn 
sqlalchemy 
pydantic 
python-jose[cryptography] 
passlib[bcrypt] 
python-multipart 
pyserial
requests
joblib 
numpy 
scikit-learn